requirements.txt
astroid==2.3.3
autopep8==1.5
Django==2.2.10
django-appconf==1.0.3
django-crispy-forms==1.9.0
django-crum==0.7.5
django-debug-toolbar==2.2
django-imagekit==4.0.2
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pilkit==2.0
Pillow==7.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
Unidecode==1.1.1
wrapt==1.11.2
